# pages/api/documentation/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blogverse API Documentation",
    "version": "1.0"
  },
  "paths": {
    "/api/blog_space": {
      "post": {
        "description": "Create a new blog space",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "The name of the blog space",
                    "example": "Mountain Biking"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The blog space was created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/blog_space"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The blog space already exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "The error message"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/500_error"
                }
              }
            }
          }
        }
      },
      "get": {
        "description": "Gets all blog spaces and their (abbreviated) posts",
        "responses": {
          "200": {
            "description": "Blog spaces with their posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/blog_space_with_posts"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "500_error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "The error message"
          }
        }
      },
      "blog_space": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the blog space"
          },
          "created_at": {
            "type": "string",
            "description": "The date the blog space was created"
          },
          "name": {
            "type": "string",
            "description": "The name of the blog space"
          },
          "slug": {
            "type": "string",
            "description": "The slug of the blog space"
          },
          "description": {
            "type": "string",
            "description": "The description of the blog space"
          },
          "about_us": {
            "type": "string",
            "description": "The about us of the blog space"
          }
        }
      },
...
      "abbreviated_post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the post"
          },
          "title": {
            "type": "string",
            "description": "The title of the post"
          },
          "slug": {
            "type": "string",
            "description": "The slug of the post"
          }
        }
      },
      "blog_space_with_posts": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the blog space"
          },
          "created_at": {
            "type": "string",
            "description": "The date the blog space was created"
          },
          "name": {
            "type": "string",
            "description": "The name of the blog space"
          },
          "slug": {
            "type": "string",
            "description": "The slug of the blog space"
          },
          "description": {
            "type": "string",
            "description": "The description of the blog space"
          },
          "about_us": {
            "type": "string",
            "description": "The about us of the blog space"
          },
          "posts": {
            "type": "array",
            "description": "The posts of the blog space",
            "items": {
              "$ref": "#/components/schemas/abbreviated_post"
            }
          }
        }
      },
...
    }
  },
...
}